backend/place-info/requirements.txt
googlemaps>=4.4.1
python-dotenv>=0.13.0